Exact matching to WCVP.

Exact matching of names to the WCVP, optionally using the author string to refine results.

exact_match(names_df, wcvp_names, name_col, author_col = NULL, id_col)

Arguments

names_df

a data frame of names for matching.

wcvp_names

A data frame of taxonomic names from WCVP version 7 or later. If NULL, names will be loaded from rWCVPdata::wcvp_names.

name_col

the column in names_df that has the taxon name for matching.

author_col

the column in names_df that has the name authority, to aid matching. Set to NULL to match with no author string.

id_col

the column in names_df that has the observation id.

Value

Match results from WCVP bound to the original data from names_df.

See also

Examples

wcvp_names <- rWCVPdata::wcvp_names

# including author string
exact_match(redlist_example, wcvp_names, "scientificName", author_col="authority",
id_col = "assessmentId")
#> # A tibble: 20 × 16
#>    assessmentId scientificName              redlistCategory authority match_type
#>           <dbl> <chr>                       <chr>           <chr>     <chr>     
#>  1     11081542 Antimima quartzitica        Least Concern   (Dinter)… NA        
#>  2     19395021 Avena hybrida               Data Deficient  Peterm.   Exact (wi…
#>  3     64135503 Citrus garrawayi            Least Concern   F.M.Bail… NA        
#>  4    189601563 Croton campanulatus         Endangered      Caruzo &… Exact (wi…
#>  5    115968141 Cynanchum freemani          Endangered      (N.E.Br.… NA        
#>  6     11047751 Echinacanthus longipes      Vulnerable      H.S.Lo &… NA        
#>  7     11001316 Geissanthus pinchinchana    Endangered      (Lundell… NA        
#>  8    126598076 Juglans pyriformis          Endangered      Liebm.    Exact (wi…
#>  9    198678856 Leichhardtia variifolia     Vulnerable      (Guillau… NA        
#> 10    135836392 Mouriri myrtilloides        Least Concern   (Sw.) Po… NA        
#> 11    146459149 Neocussonia umbellifera     Least Concern   (Sond.) … Exact (wi…
#> 12    170239556 Papuodendron lepidotum      Least Concern   C.T.White Exact (wi…
#> 13     68117888 Plerandra sp. nov. 'longis… Endangered      Lowry & … NA        
#> 14    133242195 Psammisia sclerantha        Least Concern   A.C. Sm.  NA        
#> 15       580432 Rebutia albipectinata       Endangered      Rausch    NA        
#> 16    133163169 Senecio canescens           Least Concern   (Bonpl.)… NA        
#> 17    185585004 Serruria collina            Endangered      Salisb. … NA        
#> 18      8379911 Sesbania brevipedunculata   Least Concern   J.B.Gill… NA        
#> 19    122659437 Trichilia deminuta          Critically End… D.Penn.   NA        
#> 20      1433575 Vicia mollis                Least Concern   Boiss. &… NA        
#> # … with 11 more variables: multiple_matches <lgl>, match_similarity <dbl>,
#> #   match_edit_distance <dbl>, wcvp_id <chr>, wcvp_name <chr>,
#> #   wcvp_authors <chr>, wcvp_rank <chr>, wcvp_status <chr>,
#> #   wcvp_homotypic <lgl>, wcvp_ipni_id <chr>, wcvp_accepted_id <chr>

# without author string
exact_match(redlist_example, wcvp_names, "scientificName", id_col = "assessmentId")
#> # A tibble: 23 × 16
#>    assessmentId scientificName           redlistCategory authority    match_type
#>           <dbl> <chr>                    <chr>           <chr>        <chr>     
#>  1     11081542 Antimima quartzitica     Least Concern   (Dinter) H.… NA        
#>  2     19395021 Avena hybrida            Data Deficient  Peterm.      Exact (wi…
#>  3     64135503 Citrus garrawayi         Least Concern   F.M.Bailey   NA        
#>  4    189601563 Croton campanulatus      Endangered      Caruzo & Co… Exact (wi…
#>  5    115968141 Cynanchum freemani       Endangered      (N.E.Br.) W… NA        
#>  6     11047751 Echinacanthus longipes   Vulnerable      H.S.Lo & D.… Exact (wi…
#>  7     11001316 Geissanthus pinchinchana Endangered      (Lundell) P… NA        
#>  8    126598076 Juglans pyriformis       Endangered      Liebm.       Exact (wi…
#>  9    126598076 Juglans pyriformis       Endangered      Liebm.       Exact (wi…
#> 10    198678856 Leichhardtia variifolia  Vulnerable      (Guillaumin… NA        
#> # … with 13 more rows, and 11 more variables: multiple_matches <lgl>,
#> #   match_similarity <dbl>, match_edit_distance <dbl>, wcvp_id <chr>,
#> #   wcvp_name <chr>, wcvp_authors <chr>, wcvp_rank <chr>, wcvp_status <chr>,
#> #   wcvp_homotypic <lgl>, wcvp_ipni_id <chr>, wcvp_accepted_id <chr>