Generate a species checklist from WCVP

Generate a species checklist from WCVP

generate_checklist(
  taxon = NULL,
  rank = c("species", "genus", "family", "order", "higher"),
  area = NULL,
  synonyms = TRUE,
  render.report = FALSE,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  report.filename = NULL,
  report.dir = NULL,
  report.type = c("alphabetical", "taxonomic"),
  wcvp_names = NULL,
  wcvp_distributions = NULL
)

Arguments

taxon

Character. Taxon to be included. Defaults to NULL (no taxonomic filter; all taxa).

rank

Character. One of "species", "genus", "family", "order" or "higher", giving the rank of the value/s in taxon. Must be specified unless taxon is NULL.

area

Character. One or many WGSPRD level 3 region codes. Defaults to NULL (global).

synonyms

Logical. Include synonyms in checklist (see Details)? Defaults to TRUE.

render.report

Logical. Render the checklist as a markdown report? Defaults to FALSE.

native

Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to TRUE.

introduced

Logical. Include species occurrences flagged as introduced? Defaults to TRUE.

extinct

Logical. Include species occurrences flagged as extinct? Defaults to TRUE.

location_doubtful

Logical. Include species occurrences flagged as location_doubtful? Defaults to TRUE.

report.filename

Character. Name for the HTML file. Defaults to taxon_area_type.html

report.dir

Character. Directory for the HTML file to be saved in. Must be provided by user.

report.type

Character; one of "alphabetical" or "taxonomic". Should the generated checklist be sorted alphabetically, or by taxonomic status?

wcvp_names

A data frame of taxonomic names from WCVP version 7 or later. If NULL, names will be loaded from rWCVPdata::wcvp_names.

wcvp_distributions

A data frame of distributions from WCVP version 7 or later. If NULL, distributions will be loaded from rWCVPdata::wcvp_names.

Value

Data.frame with filtered data and, if render.report=TRUE. a report HTML file.

Details

The synonyms argument can be used to limit names to those that are Accepted and Unplaced. If synonyms = TRUE then invalid, illegitimate and other non-accepted names are also included (i.e., the checklist is not limited to names for which taxon_status == "Synonym"). Two styles of checklist are supported in rWCVP - alphabetical and taxonomic. In an alphabetical checklist, all names are arranged alphabetically with accepted names in bold, and synonyms are followed by their accepted name. For a taxonomic checklist, names are grouped by their accepted names, and synonyms are listed beneath. Both types of checklist include author, publication and distribution information, though note that family headings are only supported in alphabetical checklists (due to the additional grouping requirement of the taxonomic format).

Examples

#These examples take >10 seconds to run.
if (FALSE) {
generate_checklist(taxon="Myrtaceae", rank="family", area=get_wgsrpd3_codes("Brazil"))
generate_checklist(taxon="Ferns", rank="higher", area=get_wgsrpd3_codes("New Zealand"))
}