Get area description from vector of area codes

Get area description from vector of area codes

get_area_name(area.codes)

Arguments

area.codes

Character vector containing the set of codes to be mapped to a name.

Value

Character. Either a vector of length one, with a name for the set of Level 3 areas, or (if no name exists for that set of areas) the input vector of codes.

Details

Usually used as an inverse function for get_wgsrpd3_codes. Useful for condensing sets of codes for e.g. file names, plotting and table formatting.

Examples

get_area_name(get_wgsrpd3_codes("Brazil"))
#> ℹ Matches to input geography found at Country (Gallagher) and Region (Level 2)
#> [1] "Brazil"