Generate a summary table from the WCVP

Generate a summary table from the WCVP

summary_table(
  taxon = NULL,
  rank = c("species", "genus", "family", "order", "higher"),
  area = NULL,
  grouping.var = c("area_code_l3", "genus", "family", "order", "higher"),
  wcvp_names = NULL,
  wcvp_distributions = NULL
)

Arguments

taxon

Character. Taxon to be included. Defaults to NULL (no taxonomic filter; all taxa).

rank

Character. One of "genus", "family", "order" or "higher", giving the rank of the value/s in taxon. Must be specified unless taxon is NULL.

area

Character. One or many WGSPRD level 3 region codes. Defaults to NULL (global).

grouping.var

Character; one of "area_code_l3", "genus", "family","order" or "higher" specifying how the summary should be arranged. Defaults to area_code_l3.

wcvp_names

A data frame of taxonomic names from WCVP version 7 or later. If NULL, names will be loaded from rWCVPdata::wcvp_names.

wcvp_distributions

A data frame of distributions from WCVP version 7 or later. If NULL, distributions will be loaded from rWCVPdata::wcvp_names.

Value

Data.frame with filtered data, or a gt table

Details

Valid values for rank 'higher' are 'Angiosperms', 'Gymnosperms', 'Ferns' and 'Lycophytes'.Note that grouping variable (if taxonomic) should be of a lower level than taxon and rank to produce a meaningful summary (i.e., it does not make sense to group a genus by genus, family or higher classification). Additionally, if the grouping variable is taxonomic then species occurrences are aggregated across the input area. This means that if a species is native to any of the input area (even if it is introduced or extinct in other parts) it is counted as 'Native'. Similarly, introduced occurrences take precedence over extinct occurrences. Note that in this type of summary table, 'Endemic' means endemic to the input area, not necessarily to a single WGSRPD Level 3 Area within the input area.

Examples

ferns <- summary_table("Ferns", "higher", get_wgsrpd3_codes("New Zealand"), grouping.var="family")
#> ℹ Matches to input geography found at Country (Gallagher) and Region (Level 2)
#> ℹ Aggregating occurrence types across input areas - see `?summary_table()` for details.
summary_table_gt(ferns)
#> <div id="mstucnukyn" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #mstucnukyn .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: left;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #mstucnukyn .gt_subtitle {
#>   color: #333333;
#>   font-size: 12px;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #mstucnukyn .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #mstucnukyn .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #mstucnukyn .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #mstucnukyn .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #mstucnukyn .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #mstucnukyn .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mstucnukyn .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #mstucnukyn .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #mstucnukyn .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #mstucnukyn .gt_row {
#>   padding-top: 1px;
#>   padding-bottom: 1px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: rgba(255, 255, 255, 0);
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #mstucnukyn .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mstucnukyn .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #mstucnukyn .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #mstucnukyn .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mstucnukyn .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #mstucnukyn .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mstucnukyn .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #mstucnukyn .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-left: 4px;
#>   padding-right: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mstucnukyn .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #mstucnukyn .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #mstucnukyn .gt_left {
#>   text-align: left;
#> }
#> 
#> #mstucnukyn .gt_center {
#>   text-align: center;
#> }
#> 
#> #mstucnukyn .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #mstucnukyn .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #mstucnukyn .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #mstucnukyn .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #mstucnukyn .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #mstucnukyn .gt_two_val_uncert {
#>   display: inline-block;
#>   line-height: 1em;
#>   text-align: right;
#>   font-size: 60%;
#>   vertical-align: -0.25em;
#>   margin-left: 0.1em;
#> }
#> 
#> #mstucnukyn .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#> }
#> 
#> #mstucnukyn .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #mstucnukyn .gt_slash_mark {
#>   font-size: 0.7em;
#>   line-height: 0.7em;
#>   vertical-align: 0.15em;
#> }
#> 
#> #mstucnukyn .gt_fraction_numerator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: 0.45em;
#> }
#> 
#> #mstucnukyn .gt_fraction_denominator {
#>   font-size: 0.6em;
#>   line-height: 0.6em;
#>   vertical-align: -0.05em;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="6" class="gt_heading gt_title gt_font_normal" style><b> Ferns of New Zealand </b></th>
#>     </tr>
#>     <tr>
#>       <th colspan="6" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border" style>Total number of species: 214<br>Number of regionally endemic species: 82</th>
#>     </tr>
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">Family</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Native</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Endemic</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Introduced</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Extinct</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Total</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td class="gt_row gt_left">Aspleniaceae</td>
#> <td class="gt_row gt_right">62</td>
#> <td class="gt_row gt_right">25</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">62</td></tr>
#>     <tr><td class="gt_row gt_left">Cyatheaceae</td>
#> <td class="gt_row gt_right">11</td>
#> <td class="gt_row gt_right">8</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">12</td></tr>
#>     <tr><td class="gt_row gt_left">Dennstaedtiaceae</td>
#> <td class="gt_row gt_right">12</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">12</td></tr>
#>     <tr><td class="gt_row gt_left">Gleicheniaceae</td>
#> <td class="gt_row gt_right">10</td>
#> <td class="gt_row gt_right">2</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">10</td></tr>
#>     <tr><td class="gt_row gt_left">Hymenophyllaceae</td>
#> <td class="gt_row gt_right">29</td>
#> <td class="gt_row gt_right">16</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">29</td></tr>
#>     <tr><td class="gt_row gt_left">Lindsaeaceae</td>
#> <td class="gt_row gt_right">3</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">3</td></tr>
#>     <tr><td class="gt_row gt_left">Marattiaceae</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">1</td></tr>
#>     <tr><td class="gt_row gt_left">Marsileaceae</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">1</td></tr>
#>     <tr><td class="gt_row gt_left">Ophioglossaceae</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">6</td></tr>
#>     <tr><td class="gt_row gt_left">Osmundaceae</td>
#> <td class="gt_row gt_right">4</td>
#> <td class="gt_row gt_right">3</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">5</td></tr>
#>     <tr><td class="gt_row gt_left">Polypodiaceae</td>
#> <td class="gt_row gt_right">36</td>
#> <td class="gt_row gt_right">17</td>
#> <td class="gt_row gt_right">7</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">43</td></tr>
#>     <tr><td class="gt_row gt_left">Psilotaceae</td>
#> <td class="gt_row gt_right">6</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">6</td></tr>
#>     <tr><td class="gt_row gt_left">Pteridaceae</td>
#> <td class="gt_row gt_right">16</td>
#> <td class="gt_row gt_right">5</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">17</td></tr>
#>     <tr><td class="gt_row gt_left">Salviniaceae</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">2</td></tr>
#>     <tr><td class="gt_row gt_left">Schizaeaceae</td>
#> <td class="gt_row gt_right">5</td>
#> <td class="gt_row gt_right">1</td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right"></td>
#> <td class="gt_row gt_right" style="font-weight: bold;">5</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>