Generate a species checklist from WCVP

Generate a species checklist from WCVP

Usage

wcvp_checklist(
  taxon = NULL,
  taxon_rank = c("species", "genus", "family", "order", "higher"),
  area_codes = NULL,
  synonyms = TRUE,
  render_report = FALSE,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  hybrids = FALSE,
  infraspecies = TRUE,
  report_filename = NULL,
  report_dir = NULL,
  report_type = c("alphabetical", "taxonomic"),
  wcvp_names = NULL,
  wcvp_distributions = NULL
)

Arguments

taxon

Character. Taxon to be included. Defaults to NULL (no taxonomic filter; all taxa).

taxon_rank

Character. One of "species", "genus", "family", "order" or "higher", giving the rank of the value/s in taxon. Must be specified unless taxon is NULL.

area_codes

Character. One or many WGSPRD level 3 region codes. Defaults to NULL (global).

synonyms

Logical. Include synonyms in checklist (see Details)? Defaults to TRUE.

render_report

Logical. Render the checklist as a markdown report? Defaults to FALSE.

native

Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to TRUE.

introduced

Logical. Include species occurrences flagged as introduced? Defaults to TRUE.

extinct

Logical. Include species occurrences flagged as extinct? Defaults to TRUE.

location_doubtful

Logical. Include species occurrences flagged as location_doubtful? Defaults to TRUE.

hybrids

Logical. Include hybrid species in checklist? Defaults to FALSE.

infraspecies

Logical. Include hybrid species in checklist? Defaults to TRUE.

report_filename

Character. Name for the HTML file. Defaults to taxon_area_type.html

report_dir

Character. Directory for the HTML file to be saved in. Must be provided by user.

report_type

Character; one of "alphabetical" (the default) or "taxonomic". Should the generated checklist be sorted alphabetically, or by taxonomic status?

wcvp_names

A data frame of taxonomic names from WCVP version 7 or later. If NULL (the default), names will be loaded from rWCVPdata::wcvp_names.

wcvp_distributions

A data frame of distributions from WCVP version 7 or later. If NULL (the default), distributions will be loaded from rWCVPdata::wcvp_names.

Value

Data frame with filtered data and, if render_report=TRUE. a report HTML file.

Details

The synonyms argument can be used to limit names to those that are Accepted. If synonyms = TRUE then invalid, illegitimate and other non-accepted names are also included (i.e., the checklist is not limited to names for which taxon_status == "Synonym"). Two styles of checklist are supported in rWCVP - alphabetical and taxonomic. In an alphabetical checklist, all names are arranged alphabetically with accepted names in bold, and synonyms are followed by their accepted name. For a taxonomic checklist, names are grouped by their accepted names, and synonyms are listed beneath. Both types of checklist include author, publication and distribution information, though note that family headings are only supported in alphabetical checklists (due to the additional grouping requirement of the taxonomic format).

Examples

# These examples take >10 seconds to run and require 'rWCVPdata'
# \donttest{
if(requireNamespace("rWCVPdata")){
wcvp_checklist(taxon = "Myrtaceae", taxon_rank = "family", area = get_wgsrpd3_codes("Brazil"))
wcvp_checklist(taxon = "Ferns", taxon_rank = "higher", area = get_wgsrpd3_codes("New Zealand")) %>%
head()
}
#> ℹ Matches to input geography found at Country (Gallagher) and Region (Level 2)
#> ℹ Matches to input geography found at Country (Gallagher) and Region (Level 2)
#>   higher        order       family plant_name_id    ipni_id taxon_rank
#> 1  Ferns Polypodiales Aspleniaceae       3145698   146436-3 Subspecies
#> 2  Ferns Polypodiales Aspleniaceae       3145698   146436-3 Subspecies
#> 3  Ferns Polypodiales Aspleniaceae       3145698   146436-3 Subspecies
#> 4  Ferns Polypodiales Aspleniaceae       3145698   146436-3 Subspecies
#> 5  Ferns Polypodiales Aspleniaceae       3145698   146436-3 Subspecies
#> 6  Ferns Polypodiales Aspleniaceae       3197670 17572170-1 Subspecies
#>   taxon_status genus_hybrid     genus species_hybrid        species
#> 1     Accepted         <NA> Asplenium           <NA> appendiculatum
#> 2     Accepted         <NA> Asplenium           <NA> appendiculatum
#> 3     Accepted         <NA> Asplenium           <NA> appendiculatum
#> 4     Accepted         <NA> Asplenium           <NA> appendiculatum
#> 5     Accepted         <NA> Asplenium           <NA> appendiculatum
#> 6     Accepted         <NA> Asplenium           <NA> appendiculatum
#>   infraspecific_rank   infraspecies parenthetical_author primary_author
#> 1             subsp. appendiculatum                 <NA>           <NA>
#> 2             subsp. appendiculatum                 <NA>           <NA>
#> 3             subsp. appendiculatum                 <NA>           <NA>
#> 4             subsp. appendiculatum                 <NA>           <NA>
#> 5             subsp. appendiculatum                 <NA>           <NA>
#> 6             subsp.      maritimum             Brownsey       Brownsey
#>   publication_author place_of_publication volume_and_page first_published
#> 1               <NA>              Unknown            <NA>            <NA>
#> 2               <NA>              Unknown            <NA>            <NA>
#> 3               <NA>              Unknown            <NA>            <NA>
#> 4               <NA>              Unknown            <NA>            <NA>
#> 5               <NA>              Unknown            <NA>            <NA>
#> 6               <NA>  New Zealand J. Bot.         37: 369          (1999)
#>   nomenclatural_remarks                                     geographic_area
#> 1                  <NA> SW. Victoria, Tasmania, New Zealand, Antipodean Is.
#> 2                  <NA> SW. Victoria, Tasmania, New Zealand, Antipodean Is.
#> 3                  <NA> SW. Victoria, Tasmania, New Zealand, Antipodean Is.
#> 4                  <NA> SW. Victoria, Tasmania, New Zealand, Antipodean Is.
#> 5                  <NA> SW. Victoria, Tasmania, New Zealand, Antipodean Is.
#> 6                  <NA>                                         New Zealand
#>    lifeform_description climate_description
#> 1                  <NA>                <NA>
#> 2                  <NA>                <NA>
#> 3                  <NA>                <NA>
#> 4                  <NA>                <NA>
#> 5                  <NA>                <NA>
#> 6 perennial or epiphyte           temperate
#>                                       taxon_name       taxon_authors
#> 1 Asplenium appendiculatum subsp. appendiculatum                <NA>
#> 2 Asplenium appendiculatum subsp. appendiculatum                <NA>
#> 3 Asplenium appendiculatum subsp. appendiculatum                <NA>
#> 4 Asplenium appendiculatum subsp. appendiculatum                <NA>
#> 5 Asplenium appendiculatum subsp. appendiculatum                <NA>
#> 6      Asplenium appendiculatum subsp. maritimum (Brownsey) Brownsey
#>   accepted_plant_name_id                                  accepted_name
#> 1                3145698 Asplenium appendiculatum subsp. appendiculatum
#> 2                3145698 Asplenium appendiculatum subsp. appendiculatum
#> 3                3145698 Asplenium appendiculatum subsp. appendiculatum
#> 4                3145698 Asplenium appendiculatum subsp. appendiculatum
#> 5                3145698 Asplenium appendiculatum subsp. appendiculatum
#> 6                3197670      Asplenium appendiculatum subsp. maritimum
#>   basionym_plant_name_id replaced_synonym_author homotypic_synonym
#> 1                     NA                    <NA>                NA
#> 2                     NA                    <NA>                NA
#> 3                     NA                    <NA>                NA
#> 4                     NA                    <NA>                NA
#> 5                     NA                    <NA>                NA
#> 6                3166250                    <NA>                NA
#>   parent_plant_name_id    powo_id hybrid_formula reviewed plant_locality_id
#> 1              3145697   146436-3           <NA>        N           3507645
#> 2              3145697   146436-3           <NA>        N           3507644
#> 3              3145697   146436-3           <NA>        N           3507646
#> 4              3145697   146436-3           <NA>        N           3507643
#> 5              3145697   146436-3           <NA>        N           3507647
#> 6              3145697 17572170-1           <NA>        N           4169299
#>   continent_code_l1   continent region_code_l2      region area_code_l3
#> 1                 5 AUSTRALASIA             50   Australia          TAS
#> 2                 5 AUSTRALASIA             50   Australia          VIC
#> 3                 5 AUSTRALASIA             51 New Zealand          ATP
#> 4                 5 AUSTRALASIA             51 New Zealand          NZN
#> 5                 5 AUSTRALASIA             51 New Zealand          NZS
#> 6                 5 AUSTRALASIA             51 New Zealand          NZN
#>                area introduced extinct location_doubtful occurrence_type
#> 1          Tasmania          0       0                 0          native
#> 2          Victoria          0       0                 0          native
#> 3    Antipodean Is.          0       0                 0          native
#> 4 New Zealand North          0       0                 0          native
#> 5 New Zealand South          0       0                 0          native
#> 6 New Zealand North          0       0                 0          native
#>   in_geography endemic area_endemic
#> 1        FALSE   FALSE        FALSE
#> 2        FALSE   FALSE        FALSE
#> 3         TRUE   FALSE        FALSE
#> 4         TRUE   FALSE        FALSE
#> 5         TRUE   FALSE        FALSE
#> 6         TRUE   FALSE         TRUE
# }