Reformat local versions of WCVP

Reformat local versions of WCVP

wcvp_reformat(wcvp_local, version = NULL)

Arguments

wcvp_local

Data.frame. Local copy of the WCVP.

version

Either 9 or "v9". We will add support for other versions as needed.

Value

A data.frame with the same variable structure as the WCVP that is included in the data package rWCVPdata.

Details

Note that not all of the original variables are preserved during reformatting. For example, publication is a single variable in v9, but split over multiple in the data package. It is therefore not possible to simply rename this variable. Variables that are present in the data package but not in v9 are filled with NA.