Match names to the WCVP.

Match names to WCVP, first using exact matching and then using fuzzy matching on any remaining unmatched names.

wcvp_match_names(
  names_df,
  wcvp_names = NULL,
  name_col = NULL,
  id_col = NULL,
  author_col = NULL,
  join_cols = NULL,
  fuzzy = TRUE
)

Arguments

names_df

a data frame of names for matching.

wcvp_names

A data frame of taxonomic names from WCVP version 7 or later. If NULL, names will be loaded from rWCVPdata::wcvp_names.

name_col

the column in names_df that has the taxon name for matching.

id_col

a column in names_df with a unique ID for each name. Will be created from the row number if not provided.

author_col

the column in names_df that has the name authority, to aid matching. Set to NULL to match with no author string.

join_cols

a character vector of name parts to make the taxon name, if name_col is not provided.

fuzzy

whether or not fuzzy matching should be used for names that could not be matched exactly.

Value

Match results from WCVP bound to the original data from names_df.

Details

By default, exact matching uses only the taxon name (supplied by name_col) unless a column specifying the author string is provided (as author_col).

Columns setting out name parts can be supplied as join_cols in place of a taxon name, but must be supplied in the order you want them joined (e.g. c("genus", "species", "infra_rank", "infra")).

Fuzzy matching uses a combination of phonetic and edit distance matching, and can optionally be turned off.

The WCVP can be loaded for matching from rWCVPdata::wcvp_names.

See also

Examples

wcvp_names <- rWCVPdata::wcvp_names

# without author
wcvp_match_names(redlist_example, wcvp_names, name_col="scientificName",
            id_col="assessmentId")
#> 
#> ── Matching names to WCVP ──────────────────────────────────────────────────────
#> ℹ Using the `scientificName` column
#> ! No author information supplied - matching on taxon name only
#> 
#> ── Exact matching 20 names ──
#> 
#> ✔ Found 11 of 20 names
#> 
#> ── Fuzzy matching 9 names ──
#> 
#> Matching ■■■■■■                            17% | ETA: 29s
#> Matching ■■■■■■■■■■■■■■■■                  50% | ETA: 15s
#> Matching ■■■■■■■■■■■■■■■■■■■■■             67% | ETA: 11s
#> Matching ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  5s
#> Matching ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ✔ Found 7 of 9 names
#> 
#> ── Matching complete! ──
#> 
#> ✔ Matched 18 of 20 names
#> ℹ Exact (without author): 11
#> ℹ Fuzzy (edit distance): 4
#> ℹ Fuzzy (phonetic): 3
#> ! Names with multiple matches: 3
#> # A tibble: 23 × 16
#>    assessmentId scient…¹ redli…² autho…³ match…⁴ multi…⁵ match…⁶ match…⁷ wcvp_id
#>           <dbl> <chr>    <chr>   <chr>   <chr>   <lgl>     <dbl>   <dbl> <chr>  
#>  1     11081542 Antimim… Least … (Dinte… Fuzzy … FALSE     0.95        1 642383…
#>  2     19395021 Avena h… Data D… Peterm. Exact … FALSE     1           0 397612…
#>  3     64135503 Citrus … Least … F.M.Ba… Fuzzy … FALSE     0.882       2 724094…
#>  4    189601563 Croton … Endang… Caruzo… Exact … FALSE     1           0 457017…
#>  5    115968141 Cynanch… Endang… (N.E.B… Fuzzy … FALSE     0.947       1 503650…
#>  6     11047751 Echinac… Vulner… H.S.Lo… Exact … FALSE     1           0 781525…
#>  7     11001316 Geissan… Endang… (Lunde… Fuzzy … FALSE     0.875       3 818595…
#>  8    126598076 Juglans… Endang… Liebm.  Exact … TRUE      1           0 331743…
#>  9    126598076 Juglans… Endang… Liebm.  Exact … TRUE      1           0 331744…
#> 10    198678856 Leichha… Vulner… (Guill… NA      FALSE    NA          NA NA     
#> # … with 13 more rows, 7 more variables: wcvp_name <chr>, wcvp_authors <chr>,
#> #   wcvp_rank <chr>, wcvp_status <chr>, wcvp_homotypic <lgl>,
#> #   wcvp_ipni_id <chr>, wcvp_accepted_id <chr>, and abbreviated variable names
#> #   ¹​scientificName, ²​redlistCategory, ³​authority, ⁴​match_type,
#> #   ⁵​multiple_matches, ⁶​match_similarity, ⁷​match_edit_distance

# with author
wcvp_match_names(redlist_example, wcvp_names, name_col="scientificName",
            id_col="assessmentId", author_col="authority")
#> 
#> ── Matching names to WCVP ──────────────────────────────────────────────────────
#> ℹ Using the `scientificName` column
#> 
#> ── Exact matching 20 names ──
#> 
#> ✔ Found 11 of 15 names
#> 
#> ── Fuzzy matching 9 names ──
#> 
#> Matching ■■■■■■                            17% | ETA: 18s
#> Matching ■■■■■■■■■■■■■■■■                  50% | ETA: 15s
#> Matching ■■■■■■■■■■■■■■■■■■■■■             67% | ETA: 11s
#> Matching ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  5s
#> Matching ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ✔ Found 7 of 9 names
#> 
#> ── Matching complete! ──
#> 
#> ✔ Matched 18 of 20 names
#> ℹ Exact (with author): 5
#> ℹ Exact (without author): 6
#> ℹ Fuzzy (edit distance): 4
#> ℹ Fuzzy (phonetic): 3
#> ! Names with multiple matches: 2
#> # A tibble: 22 × 18
#>    assessmentId scient…¹ redli…² autho…³ match…⁴ multi…⁵ match…⁶ match…⁷ wcvp_id
#>           <dbl> <chr>    <chr>   <chr>   <chr>   <lgl>     <dbl>   <dbl> <chr>  
#>  1     11081542 Antimim… Least … (Dinte… Fuzzy … FALSE     0.95        1 642383…
#>  2     19395021 Avena h… Data D… Peterm. Exact … FALSE     1           0 397612…
#>  3     64135503 Citrus … Least … F.M.Ba… Fuzzy … FALSE     0.882       2 724094…
#>  4    189601563 Croton … Endang… Caruzo… Exact … FALSE     1           0 457017…
#>  5    115968141 Cynanch… Endang… (N.E.B… Fuzzy … FALSE     0.947       1 503650…
#>  6     11047751 Echinac… Vulner… H.S.Lo… Exact … FALSE     1           0 781525…
#>  7     11001316 Geissan… Endang… (Lunde… Fuzzy … FALSE     0.875       3 818595…
#>  8    126598076 Juglans… Endang… Liebm.  Exact … FALSE     1           0 331743…
#>  9    198678856 Leichha… Vulner… (Guill… NA      FALSE    NA          NA NA     
#> 10    135836392 Mouriri… Least … (Sw.) … Exact … FALSE     1           0 101815…
#> # … with 12 more rows, 9 more variables: wcvp_name <chr>, wcvp_authors <chr>,
#> #   wcvp_rank <chr>, wcvp_status <chr>, wcvp_homotypic <lgl>,
#> #   wcvp_ipni_id <chr>, wcvp_accepted_id <chr>,
#> #   wcvp_author_edit_distance <dbl>, wcvp_author_lcs <int>, and abbreviated
#> #   variable names ¹​scientificName, ²​redlistCategory, ³​authority, ⁴​match_type,
#> #   ⁵​multiple_matches, ⁶​match_similarity, ⁷​match_edit_distance